Update to JDK 11, NetBeans 11.3 and GeoTools 28.2 (#18)
* Snap 3454 tb geotools (#15)

* Migration to GeoTools 28

* GeoTools update to 28.2

* intermediate - loading issues

* replaced commons-lang with commons-lang3

* updated mockito

* further updates for JDK11

---------


* set Java 11 a source and target version of compilation

* updated mockito

* Using NetBeans 10

* set version to 10.0.0-JDK11SHOT to differentiate from master

* changes groupId of seadas modules

* changed version for seadas modules

* updated installer

* consider -JDK11SHOT in build helper and fixed seadas groupid

* Snap 3476 opttbx s2msi reader fix for JDK11 branch (#16)

* (opttbx-s2msi-reader): Revert this solution to include it on the corresponding JIRA.

* [SNAP-3476](opttbx-s2msi-reader): Fix error opening bands from Sen2Cor output products.

---------


* Snap 3460 opttbx sta adapters sen2cor211 for JDK 11 (#17)

* [SNAP-3460](opttbx-sta-adapters_sen2cor211): STA adapter for Sen2Cor 2.11.

* [SNAP-3460](opttbx-sta-adapters_sen2cor211): update STA adapter for Sen2Cor 2.11.

* [SNAP-3460](opttbx-sta-adapters_sen2cor211): update descriptor

* [SNAP-3460](opttbx-sta-adapters_sen2cor211): update adapter

* [SNAP-3460](opttbx-sta-adapters_sen2cor211): update STA adapter for Sen2Cor 2.11: Add missing parameter for output directory

* [SNAP-3460](opttbx-sta-adapters_sen2cor211): Fix problem when invalid input was provided or the output product not written.

* [SNAP-3460](opttbx-sta-adapters_sen2cor211): Add missing parameters and templates to fix problem with missing of band B1 from output product.

* [SNAP-3460](opttbx-sta-adapters_sen2cor211): Add missing help page for Sen2Cor 2.11.

* [SNAP-3460](opttbx-sta-adapters_sen2cor211): Refactoring name of variables for pretty showing on operator UI.

* [SNAP-3460](opttbx-sta-adapters_sen2cor211): Refactoring help page for Sen2Cor 2.11.

* [SNAP-3460](opttbx-sta-adapters_sen2cor211): Update descriptor of Sen2Cor 2.1.1 for showing XML parameters on Graph Builder UI.

* [SNAP-3460](opttbx-sta-adapters_sen2cor211): UI misspellings fix.

* [SNAP-3460](opttbx-sta-adapters_sen2cor211): Help page update.

* [SNAP-3460](opttbx-sta-adapters_sen2cor211): Fix problem with target product opening in SNAP after Sen2Cor execution when outputDir parameter was specified.

* [SNAP-3460](opttbx-sta-adapters_sen2cor211): Allow selecting all resolutions for processing

---------


* Update NetBeans to 11.3

* fixed helpset for alos-reader module

* require Java > 11 in manifest.mf

* finally changed last versions to netbeans

* removed 1.8 source/target settings and set them to 11

* removed 1.8 source/target settings and set them to 11

* updated projects to make use of new help system

* updated projects to make use of new help system

---------

+++ opttbx-rapideye-reader/src/main/java/nitf/imageio/ImageIOUtils.java
@@ -492,7 +492,7 @@ public class ImageIOUtils {
      * @return An unsigned integer
      */
     public static int makeUnsigned(byte b) {
-        return b >= 0 ? (int) b : 255 + (int) b + 1;
+        return b >= 0 ? b : 255 + b + 1;
     }
 
     public static void main(String[] args) {
